fastapi-backend/requirements.txt
# FastAPI and server
fastapi==0.115.5
uvicorn[standard]==0.32.1
python-multipart==0.0.18

# Document parsing
pypdf==5.1.0
python-docx==1.1.2

# AI integration
anthropic==0.39.0

# Data validation
pydantic==2.10.3
pydantic-settings==2.6.1

# Utilities
python-dotenv==1.0.1
